# [PERF] mrp: speed up `onchange` on MO's `lot_producing_id`