@pivi-odoo 2024-06-26
Let me know if you have better suggestion for naming these variables, their name is irrelevant for effectiveness of the patch.