@Xavier-Do 2024-08-05
Upgrade exception [#369](https://runbot.odoo.com/web/#id=369&view_type=form&model=runbot.upgrade.exception) added.

Waiting the forward-port of this pr, the exception will be applied on all builds, and create an inconsistent state for migrations.
**Please forward port this asap up to master without change.** Exception should be forward-ported before the end of the week.

PRs adding fields/models should be merged at the beginning of the week and all the forward ports should be merged in the same week.

If you need to apply any change before it reaches master, please notify runbot team.

Details:
```
field:mrp.production.all_move_ids
field:mrp.production.all_move_raw_ids
```
cc @KangOl @nseinlet @aj-fuentes