@MiquelRForgeFlow 2024-08-06
There are many "missing" One2many fields in Odoo which would be "nice" if they could be added, like https://github.com/odoo/odoo/pull/134162.